chore: refactor and move a few files around
