+++ packages/insomnia-app/app/network/grpc/proto-loader/index.js
@@ -1,8 +1,8 @@
 // @flow
 
-import type { GrpcMethodDefinition } from './method';
+import type { GrpcMethodDefinition } from '../method';
 import * as protoLoader from '@grpc/proto-loader';
-import * as models from '../../models';
+import * as models from '../../../models';
 import { writeProtoFile } from './write-proto-file';
 
 const GRPC_LOADER_OPTIONS = {
