+++ packages/insomnia-app/app/network/grpc/proto-manager/ingest-proto-directory.js
@@ -1,7 +1,7 @@
 // @flow
 
-import * as models from '../models';
-import type { ProtoDirectory } from '../models/proto-directory';
+import * as models from '../../../models';
+import type { ProtoDirectory } from '../../../models/proto-directory';
 import path from 'path';
 import fs from 'fs';
